requirements.txt
Cython==0.29.13
async-timeout==3.0.0
asyncpg==0.19.0
cffi==1.12.2
chardet==3.0.4
jsonschema==2.6.0
multidict==4.5.2
pycparser==2.18
pycryptodome==3.6.6
PyJWT==1.6.0
python-dateutil==2.6.1
PyYaml==5.1
six==1.11.0
ujson==1.35
zope.interface==4.6.0
uvicorn==0.10.3
argon2-cffi==18.3.0
